+++ buildroot/packages/configd/Makefile
@@ -1,9 +1,12 @@
 CC=gcc
+CFLAGS=-I$(PWD)/package/pd690xx -I$(PWD)/package/postmerkos -Os
+LDFLAGS=-ljson-c -lwebsockets
 
-all: *.c
-	$(CC) -I$(PWD)/package/pd690xx -I$(PWD)/package/postmerkos -Os -o configd -ljson-c -lwebsockets '$<' \
-	$(PWD)/package/pd690xx/libpd690xx.c \
-	$(PWD)/package/postmerkos/libpostmerkos.c
+SRCS=main.c config.c status.c websocket.c
+LIBS=$(PWD)/package/pd690xx/libpd690xx.c $(PWD)/package/postmerkos/libpostmerkos.c
+
+all: $(SRCS)
+	$(CC) $(CFLAGS) -o configd $(SRCS) $(LIBS) $(LDFLAGS)
+
 clean:
-	rm -rf a.out cfg
-	rm -f *.o
+	rm -f configd *.o
