+++ buildroot/packages/configd/status.h
@@ -1,0 +1,8 @@
+#ifndef STATUS_H
+#define STATUS_H
+
+#include <json-c/json.h>
+
+struct json_object *get_status(void);
+
+#endif
